# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Stock and sales management API for local shop owners.",
        "title": "shopkeeper API",
        "contact": {
            "name": "Humberto",
            "url": "https://github.com/humberto0/shopkeeper"
        },
        "license": {
            "name": "MIT",
            "url": "https://github.com/humberto0/shopkeeper/blob/main/LICENSE"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/health": {
            "get": {
                "description": "Pings the database pool to check that the API and its dependencies are up.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "health"
                ],
                "summary": "Health check",
                "responses": {
                    "200": {
                        "description": "status: ok",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "503": {
                        "description": "database unavailable",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/users": {
            "post": {
                "description": "Creates a new user account (owner or clerk).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Register a new user",
                "parameters": [
                    {
                        "description": "User to register",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.registerUserRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/handler.registerUserResponse"
                        }
                    },
                    "400": {
                        "description": "malformed request body",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    },
                    "409": {
                        "description": "email already exists",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    },
                    "422": {
                        "description": "invalid name, email, password or role",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    }
                }
            }
        },
        "/users/{id}": {
            "get": {
                "description": "Find user by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Find User",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.findUserResponse"
                        }
                    },
                    "400": {
                        "description": "malformed id",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    },
                    "404": {
                        "description": "user not found",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Edit user account (owner or clerk). Version must match the current stored version (optimistic locking).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Edit an existing user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "User fields to update",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.editUserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.editUserResponse"
                        }
                    },
                    "400": {
                        "description": "malformed request body or invalid id",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    },
                    "404": {
                        "description": "user not found",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    },
                    "409": {
                        "description": "email already exists",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    },
                    "422": {
                        "description": "invalid name, email, role, or stale version",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handler.editUserRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "humberto@shop.com"
                },
                "name": {
                    "type": "string",
                    "example": "Humberto test"
                },
                "role": {
                    "type": "string",
                    "example": "owner"
                },
                "version": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "handler.editUserResponse": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "humberto@shop.com"
                },
                "id": {
                    "type": "string",
                    "example": "018f1d3a-7c3e-7c3e-8b3e-7c3e7c3e7c3e"
                },
                "name": {
                    "type": "string",
                    "example": "Humberto test"
                },
                "role": {
                    "type": "string",
                    "example": "owner"
                },
                "version": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "handler.errorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "handler.findUserResponse": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string",
                    "example": "2020-01-01T00:00:00Z"
                },
                "email": {
                    "type": "string",
                    "example": "humbertotest@shop.com"
                },
                "id": {
                    "type": "string",
                    "example": "018f1d3a-7c3e-7c3e-8b3e-7c3e7c3e7c3e"
                },
                "isActive": {
                    "type": "boolean",
                    "example": true
                },
                "name": {
                    "type": "string",
                    "example": "Humberto test"
                },
                "role": {
                    "type": "string",
                    "example": "owner"
                },
                "updatedAt": {
                    "type": "string",
                    "example": "2020-01-01T00:00:00Z"
                },
                "version": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "handler.registerUserRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "humbertotest@shop.com"
                },
                "name": {
                    "type": "string",
                    "example": "Humberto test"
                },
                "password": {
                    "type": "string",
                    "example": "senha1234"
                },
                "role": {
                    "type": "string",
                    "enum": [
                        "owner",
                        "clerk"
                    ],
                    "example": "owner"
                }
            }
        },
        "handler.registerUserResponse": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "humbertotest@shop.com"
                },
                "id": {
                    "type": "string",
                    "example": "018f1d3a-7c3e-7c3e-8b3e-7c3e7c3e7c3e"
                },
                "name": {
                    "type": "string",
                    "example": "Humberto test"
                },
                "role": {
                    "type": "string",
                    "example": "owner"
                }
            }
        }
    }
}